OU detection

+++ Private/New-OUStructure.ps1
@@ -7,6 +7,7 @@
         [Parameter(Mandatory)][string] $DC
     )
     $IgnoredProperties = @('Delegation', 'DelegationInheritance', 'CanonicalNameOU')
+    $OUProperties = @('Description', 'ProtectedFromAccidentalDeletion')
     $PartsOU = $CanonicalNameOU.Split("\")
 
     $LevelPath = $BasePath
@@ -16,34 +17,39 @@
         $CanonicalCurrentPath = ConvertFrom-DistinguishedName -DistinguishedName ($CurrentPath.Replace(",$BasePath", "")) -ToCanonicalName
 
         # lets create new OU, but if it exists we will just inform user
-        try {
-            $newADOrganizationalUnitSplat = @{
-                Name                            = $O
-                Path                            = $LevelPath
-                ProtectedFromAccidentalDeletion = $False
-                Server                          = $DC
-                ErrorAction                     = 'Stop'
-            }
-            if ($i -eq ($PartsOU.Count - 1)) {
-                foreach ($V in $ConfigurationOU.Keys) {
-                    if ($V -notin $IgnoredProperties) {
-                        $newADOrganizationalUnitSplat[$V] = $ConfigurationOU[$V]
+        $OrganizationalUnitExists = Get-ADOrganizationalUnit -Filter "DistinguishedName -eq '$CurrentPath'" -ErrorAction SilentlyContinue -Properties $OUProperties -Server $DC
+        if (-not $OrganizationalUnitExists) {
+            try {
+                $newADOrganizationalUnitSplat = @{
+                    Name                            = $O
+                    Path                            = $LevelPath
+                    ProtectedFromAccidentalDeletion = $False
+                    Server                          = $DC
+                    ErrorAction                     = 'Stop'
+                }
+                if ($i -eq ($PartsOU.Count - 1)) {
+                    foreach ($V in $ConfigurationOU.Keys) {
+                        if ($V -notin $IgnoredProperties) {
+                            $newADOrganizationalUnitSplat[$V] = $ConfigurationOU[$V]
+                        }
                     }
                 }
+                Remove-EmptyValue -Hashtable $newADOrganizationalUnitSplat
+                if ($newADOrganizationalUnitSplat.Count -eq 1) {
+                    #Write-Color -Text '[!] ', "No OU will be created ", $CurrentPath -Color Red, White
+                } else {
+                    New-ADOrganizationalUnit @newADOrganizationalUnitSplat
+                    Write-Color -Text '[+]', "[$CanonicalCurrentPath]", "[Adding]", " Added new organizational unit" -Color Green, DarkGray, Green, White
+                }
+            } catch {
+                if ($_.Exception.Message -notlike '*with a name that is already in use*') {
+                    Write-Color -Text '[!]', "[$CanonicalCurrentPath]", "[Error]", " Error $($_.Exception.message)" -Color Red, DarkGray, Yellow, DarkMagenta, White
+                } else {
+                    Write-Color -Text '[*]', "[$CanonicalCurrentPath]", "[Skipping]", " Skipped new organizational unit, already exists!" -Color DarkMagenta, DarkGray, Yellow, DarkMagenta, White
+                }
             }
-            Remove-EmptyValue -Hashtable $newADOrganizationalUnitSplat
-            if ($newADOrganizationalUnitSplat.Count -eq 1) {
-                #Write-Color -Text '[!] ', "No OU will be created ", $CurrentPath -Color Red, White
-            } else {
-                New-ADOrganizationalUnit @newADOrganizationalUnitSplat
-                Write-Color -Text '[+]', "[$CanonicalCurrentPath]", "[Adding]", " Added new organizational unit" -Color Green, DarkGray, Green, White
-            }
-        } catch {
-            if ($_.Exception.Message -notlike '*with a name that is already in use*') {
-                Write-Color -Text '[!]', "[$CanonicalCurrentPath]", "[Error]", " Error $($_.Exception.message)" -Color Red, DarkGray, Yellow, DarkMagenta, White
-            } else {
-                Write-Color -Text '[*]', "[$CanonicalCurrentPath]", "[Skipping]", " Skipped new organizational unit, already exists!" -Color DarkMagenta, DarkGray, Yellow, DarkMagenta, White
-            }
+        } else {
+            Write-Color -Text '[*]', "[$CanonicalCurrentPath]", "[Skipping]", " Skipped new organizational unit, already exists!" -Color DarkMagenta, DarkGray, Yellow, DarkMagenta, White
         }
         # once the OU is created we should update it with additional fields
         $setADOrganizationalUnitSplat = @{
@@ -54,8 +60,10 @@
         $PropertiesToUpdate = if ($i -eq ($PartsOU.Count - 1)) {
             foreach ($V in $ConfigurationOU.Keys) {
                 if ($V -notin $IgnoredProperties) {
-                    $V
-                    $setADOrganizationalUnitSplat[$V] = $ConfigurationOU[$V]
+                    if ($OrganizationalUnitExists.$V -ne $ConfigurationOU[$V]) {
+                        $V
+                        $setADOrganizationalUnitSplat[$V] = $ConfigurationOU[$V]
+                    }
                 }
             }
         }
